+++ app/src/main/java/dev/hexhydra/HooksLocation.kt
@@ -1,4 +1,4 @@
-package dev.codex.deviceprivy
+package dev.hexhydra
 
 import android.location.Location
 import android.os.Build